Remove cached artifacts and add gitignore
