+++ main.py
@@ -1,0 +1,7 @@
+"""執行入口。"""
+
+from finmind_etl.cli import main
+
+
+if __name__ == "__main__":
+    main()
